Cleaning up test code with proper asserts
